Capture cardIdentifier response from a submitted dropin form.
The solution here feels a little cludgy. We need a cardIdentifie from the
dropin form, but will only be given a card-identifier (different name - no
idea why) in the main payment form. We may also get three error parameters
instead, detailing the error code, http code and error message from the
gateway call trying to fetch the card identifier. All of these will be in
the _POST array or the body of the ServerRequest message, so we need to
be able to extract that. These changes attempt to do that, and seem to work
using `Factory\ResponseFactory::parse($_POST)`. Will revisit - maybe force
the use of the ServerRequest object and not support the arrays in the
factory response creation.