Issue #45 Renamed requests: SessionKey and SessionKeyValidate to Create-- and Fetch--
It has struck me, after far too long, that these are messages to *DO* something
and not objects that represent a resource (as the response messages mostly are).
So these messages should be named with a verb and the name of the resource they
are affecting or fetching. This feels much clearer.